Доделана страница для КЛИЕНТА

+++ EducationalPracticeAutumn2024/DB/Service.cs
@@ -23,7 +23,7 @@ namespace EducationalPracticeAutumn2024.DB
     
         public int ID { get; set; }
         public string Title { get; set; }
-        public decimal Cost { get; set; }
+        public int Cost { get; set; }
         public int DurationInSeconds { get; set; }
         public string Description { get; set; }
         public Nullable<double> Discount { get; set; }
